♻️ refactor(utils): extract resolve_model_name function to models module
- move resolve_model_name function from utils to a separate models module for better organization
- remove unused imports in utils module
